Added Reid's Data
File to CSV Functionality. OSM Program Filter

+++ InrixNode.java
@@ -1,21 +1,21 @@
 
 public class InrixNode {
 	public int type;
-	public float time;
-	public float lat;
-	public float lon;
+	public double time;
+	public double lat;
+	public double lon;
 	public int source;
-	public float ID;
+	public double ID;
 	int acData;
 	public InrixNode next;
 	public InrixNode previous;
 	public InrixNode(String[] data) {
 		this.type = Integer.parseInt(data[0]);
-		this.time = Float.parseFloat(data[1]);
-		this.lat = Float.parseFloat(data[2]);
-		this.lon = Float.parseFloat(data[3]);
+		this.time = Double.parseDouble(data[1]);
+		this.lat = Double.parseDouble(data[2]);
+		this.lon =  Double.parseDouble(data[3]);
 		this.source = Integer.parseInt(data[4]);
-		this.ID = Float.parseFloat(data[5]);
+		this.ID =  Double.parseDouble(data[5]);
 		this.acData = Integer.parseInt(data[6]);
 	}
 	public String toString() {
